+++ src/smax-util.c
@@ -4,8 +4,7 @@
  * \date Jun 25, 2019
  * \author Attila Kovacs
  *
- * \brief
- *      A collection of commonly used functions for the SMA-X library.
+ *  A collection of commonly used functions for the SMA-X library.
  */
 
 #define _GNU_SOURCE               ///< for strcasecmp()
@@ -34,7 +33,7 @@ static void *SMAXReconnectThread(void *arg);
 // Local variables ------------------------------------>
 
 /// A lock for ensuring exclusive access for pipeline configuration changes...
-/// and the variables that it controls, e.g. via lockConfig()
+/// and the variables that it controls, e.g. via smaxLockConfig() / smaxUnlockConfig()
 static pthread_mutex_t configLock = PTHREAD_MUTEX_INITIALIZER;
 
 static XBoolean isDisabled = FALSE;
@@ -52,7 +51,7 @@ int smaxLockConfig() {
 }
 
 /**
- * Release the exclusive lock to SMA-X configuration, so that others may access/update them also.
+ * Release the exclusive lock to SMA-X configuration, so that others may access/update it also.
  *
  * \return      The result of pthread_mutex_unlock().
  */
@@ -65,7 +64,7 @@ int smaxUnlockConfig() {
 
 /**
  * Creates a new SMA-X metadata object with defaults. Effectively the same as calling
- * calloc() followed by xResetMeta().
+ * malloc() followed by xResetMeta().
  *
  * @return              Pointer to a new metadata object initialized to defaults.
  *
@@ -172,7 +171,7 @@ void smaxSocketErrorHandler(Redis *redis, enum redisx_channel channel, const cha
 }
 
 /**
- * Same as smaxScriptError(), but it assumed the aller already has esclusive access to the configuration
+ * Same as smaxScriptError(), but it assumed the caller already has exclusive access to the configuration
  * state, by having called smaxConfigLock() before.
  *
  * @param name          The name of the calling function or name of script (whichever is more informative).
@@ -244,13 +243,13 @@ int smaxScriptError(const char *name, int status) {
 int smaxError(const char *func, int errorCode) {
 #if SMAX_LEGACY
   if(errorCode <= X_LEGACY_CODES) {
-    if(xDebug) fprintf(stderr, "DEBUG-X> %4d (%s) in %s.\n", errorCode, smaxErrorDescription(errorCode), func);
+    xdprintf("DEBUG-X> %4d (%s) in %s.\n", errorCode, smaxErrorDescription(errorCode), func);
     return errorCode;
   }
 #endif
 
   // If in the process of reconnecting, we don't want to spam about transmit errors here, so just return the error code...
-  if(errorCode == X_NO_SERVICE) if(isDisabled) return errorCode;
+  if(errorCode == X_NO_SERVICE && isDisabled) return errorCode;
 
   return redisxError(func, errorCode);
 }
@@ -478,7 +477,6 @@ int smaxTimestamp(char *buf) {
  *                      X_NULL          if there was no timestamp (NULL or empty string), or the `secs` argument is NULL.
  *                      X_PARSE_ERROR   if the seconds could not be parsed.
  *                      1               if there was an error parsing the nanosec part.
- *                      X_NULL          if the secs argument is NULL
  */
 int smaxParseTime(const char *timestamp, time_t *secs, long *nanosecs) {
   static const char *fn = "smaxParseTime";
@@ -506,7 +504,7 @@ int smaxParseTime(const char *timestamp, time_t *secs, long *nanosecs) {
       if(nanosecs) *nanosecs = 0;
       return 1;
     }
-    if(nanosecs) *nanosecs = (int) (1e9 * d);
+    if(nanosecs) *nanosecs = (long) (1e9 * d);
   }
   else if(nanosecs) *nanosecs = 0;
 
@@ -810,7 +808,7 @@ char *smaxValuesToString(const void *value, XType type, int eCount, char *trybuf
     x_error(0, EINVAL, fn, "structures not allowed");
     return NULL;                     // structs are not serialized by this function.
   }
-  if(type == X_RAW) if(value) return *(char **) value;
+  if(type == X_RAW) return *(char **) value;
 
   // Figure out how big the serialized string might be...
   if(type == X_UNKNOWN) stringSize = 2 * eCount;
@@ -1098,7 +1096,7 @@ int smaxStringToValues(const char *str, void *value, XType type, int eCount, int
     }
 
     // Zero out the remaining elements...
-    if(k < eCount) xZero(&c[k], type, eCount - k);
+    if(k < eCount) xZero(&c[k * eSize], type, eCount - k);
   }
 
   *pos = next - str;