+++ tests/controlTest.c
@@ -4,10 +4,9 @@
  * @date Created on: May 6, 2020
  * @author Attila Kovacs
  *
- *      This simple program demonstrates and tests the use of lazy pulling from the SMA-X database.
- *      Lazy polling allows frequent checking on (i.e. polling) an infrequently changing variable's content
- *      without causing excessive network traffic. Data is pulled from SMA-X only on the first call
- *      to smaxLazyPull(), and then only when an update notification is received for the lazy value.
+ *      This simple program demonstrates and tests the use of remote program control using SMA-X.
+ *      You may use SMA-X to control programs remotely and receive confirmation of the controlled
+ *      commands / variables.
  */
 
 #define _POSIX_C_SOURCE 199309L       ///< for nanosleep()
@@ -22,7 +21,7 @@
 #include "smax.h"
 
 #ifndef SMAX_TEST_TIMEOUT
-#  define SMAX_TEST_TIMEOUT 3     ///< [s] Default timeout
+#  define SMAX_TEST_TIMEOUT 10     ///< [s] Default timeout
 #endif
 
 #define TABLE           "_test_" X_SEP "control"
@@ -31,20 +30,25 @@
 
 
 
-static void checkStatus(char *op, int status) {
+static void checkStatus(const char *op, int status) {
   if(!status) return;
   fprintf(stderr, "ERROR! %s: %s\n", op, smaxErrorDescription(status));
-  exit(-1);
+  exit(EXIT_FAILURE);
 }
 
 int ControlFunction(const char *table, const char *key, void *parg) {
   const char *replyKey = (const char *) parg;
-  int value = smaxPullInt(table, key, -1);
+  int value = smaxPullInt(table, key, X_FAILURE);
   return smaxShareInt(table, replyKey, value);
 }
 
 int main() {
   int reply;
+
+  // Because this process is doing both the control and the service parts, we'll use the interactive
+  // channel exclusively, to avoid any race conditions on the Redis server. Normally it is not
+  // necessary when we control one process from another.
+  smaxSetPipelined(FALSE);
 
   xSetDebug(TRUE);
   //redisxDebugTraffic(TRUE);
@@ -54,17 +58,17 @@ int main() {
   // Initialize the value that we will control, and change at some later time...
   checkStatus("share", smaxShareInt(TABLE, CONTROL_NAME, 0));
 
-  checkStatus("setControlCall", smaxSetControlFunction(TABLE, CONTROL_NAME, ControlFunction, NAME));
+  checkStatus("setControlFunction", smaxSetControlFunction(TABLE, CONTROL_NAME, ControlFunction, NAME));
 
   // We'll update the value here...
   // The waiting thread should set gotUpdate when it unblocks...
   errno = 0;
-  reply = smaxControlInt(TABLE, CONTROL_NAME, 1, NULL, NAME, -1, SMAX_TEST_TIMEOUT);
+  reply = smaxControlInt(TABLE, CONTROL_NAME, 1, NULL, NAME, X_FAILURE, SMAX_TEST_TIMEOUT);
   if(reply != 1) {
     fprintf(stderr, "ERROR! Unexpected reply: expected %d, got %d.\n", 1, reply);
     if(errno) fprintf(stderr, "      errno = %d (%s)\n", errno, strerror(errno));
     fprintf(stderr, "control: FAILED\n");
-    return -1;
+    return EXIT_FAILURE;
   }
 
   fprintf(stderr, "control: OK\n");