+++ tests/simpleIntTest.c
@@ -19,7 +19,7 @@
 #define NAME    "integer"
 #define VALUE   2020
 
-static void checkStatus(char *op, int status) {
+static void checkStatus(const char *op, int status) {
   if(status >= 0) return;
   fprintf(stderr, "ERROR! %s: %s\n", op, smaxErrorDescription(status));
   exit(-1);